+++ openjaus/trunk/OpenJAUSv3.3/libjaus/src/message/query/communications/queryDataLinkStatusMessage.c
@@ -122,7 +122,7 @@ static int dataToString(QueryDataLinkStatusMessage message, char **buf)
   strcpy((*buf), "\nData Link Id: " );
   jausByteToString(message->dataLinkId, (*buf)+strlen(*buf));
   
-  return strlen((*buf));
+  return (int)strlen(*buf);
 }
 
 // Returns number of bytes put into the buffer
@@ -516,6 +516,6 @@ static int headerToString(QueryDataLinkStatusMessage message, char **buf)
   strcat((*buf), "\nSequence Number: ");
   jausUnsignedShortToString(message->sequenceNumber, (*buf)+strlen(*buf));
   
-  return strlen((*buf));
+  return (int)strlen(*buf);
   
 }
