+++ openjaus/trunk/OpenJAUSv3.3/libjaus/src/message/query/planning/querySpoolingPreferenceMessage.c
@@ -505,6 +505,6 @@ static int headerToString(QuerySpoolingPreferenceMessage message, char **buf)
   strcat((*buf), "\nSequence Number: ");
   jausUnsignedShortToString(message->sequenceNumber, (*buf)+strlen(*buf));
   
-  return strlen((*buf));
+  return (int)strlen(*buf);
   
 }
