+++ openjaus/trunk/OpenJAUSv3.3/libjaus/src/message/query/platform/queryDiscreteDevicesMessage.c
@@ -116,7 +116,7 @@ static int dataToString(QueryDiscreteDevicesMessage message, char **buf)
   strcpy((*buf), "\nPresence Vector: " );
   jausByteToHexString(message->presenceVector, (*buf)+strlen(*buf));
   
-  return strlen((*buf));
+  return (int)strlen(*buf);
 }
 
 static unsigned int dataSize(QueryDiscreteDevicesMessage message)
@@ -509,6 +509,6 @@ static int headerToString(QueryDiscreteDevicesMessage message, char **buf)
   strcat((*buf), "\nSequence Number: ");
   jausUnsignedShortToString(message->sequenceNumber, (*buf)+strlen(*buf));
   
-  return strlen((*buf));
+  return (int)strlen(*buf);
   
 }
