+++ openjaus/trunk/OpenJAUSv3.3/libjaus/src/message/query/platform/queryPathSegmentCountMessage.c
@@ -488,6 +488,6 @@ static int headerToString(QueryPathSegmentCountMessage message, char **buf)
   strcat((*buf), "\nSequence Number: ");
   jausUnsignedShortToString(message->sequenceNumber, (*buf)+strlen(*buf));
   
-  return strlen((*buf));
+  return (int)strlen(*buf);
   
 }
